% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{Billonage-package}
\alias{Billonage}
\alias{Billonage-package}
\title{Billonage: Équations de billonnage PETRO}
\description{
Équations pour estimer le volume des produits PETRO de 6 essences de feuillus. Équations PETRO 2015 ou équations PETRO régionalisées 2024.
}
\keyword{internal}
