% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIM_Billonnage_ABCD_DHP.R
\name{SIMBillonnageABCD_DHP}
\alias{SIMBillonnageABCD_DHP}
\title{Fonction prévoit la répartition par produits des arbres de 6 essences feuillus à l'aide des nouvelles
équations de Petro régionalisés ou celles de 2015.}
\usage{
SIMBillonnageABCD_DHP(Data, type)
}
\arguments{
\item{Data}{Un dataframe qui contient en ligne les arbres dont on veut prévoir
les rendements en produit Petro.
Le dataframe doit contenir une colonne bilonID qui numérote individuellement chacune des lignes.
Doit aussi contenir les colonnes Espece et DHPcm
Pour type=DHP et type=ABCD, il faut la colonne eco ou reg_eco
Pour type=ABCD ou ABCD2015, il faut aussi une colonne ABCD
Pour type=MSCR, il faut aussi une colonne MSCR
Pour type=1234, il faut aussi les colonnes prod0 et vigu0
Les équations ne s'appliquent qu'aux arbres avec un dhp>23, les autres seront supprimés
Les équations ne s'appliquent qu'aux espèces: "ERS", "BOJ", "ERR", "BOP", "HEG", "CHX", les autres seront supprimés}

\item{type}{"DHP" pour utiliser les équations régionalisées basées seulement sur le DHP
"ABCD" pour utiliser les équations régionalisées basées sur ABCD
"1234" pour utiliser les équations de 2015 basées sur 1234
"MSCR" pour utiliser les équations de 2015 basées sur MSCR
"DHP2015" pour utiliser les équations de 2015 basées seulement sur le DHP
"ABCD2015" pour utiliser les équations de 2015 basées sur ABCD}
}
\value{
Retourne un dataframe avec l'estimation du volume par classe de produit
         pour chacun des arbres "ERS", "BOJ", "ERR", "BOP", "HEG", "CHX" pour Petro2015 et ERS/BOJ pour Petro2024
}
\description{
Fonction prévoit la répartition par produits des arbres de 6 essences feuillus à l'aide des nouvelles
équations de Petro régionalisés ou celles de 2015.
}
\examples{
vol_billon <- SIMBillonnageABCD_DHP(Data=liste_arbres_ex, type="ABCD")
}
